LIBSEARCH-120. Modified environment banner to support custom colors
Moved default configuration and display logic out of the
"umd_environment_banner_helper.rb" file into
lib/quick_search-umd_theme.rb, as suggested in

https://rarlindseysmash.com/posts/config-and-generators-in-gems

This has the advantage that the configuration and display logic will
only run once per application start, minimizing the amount of time it
takes to generate the banner.

The banner can be configured using the following (optional)
environment variables:

* ENVIRONMENT_BANNER
* ENVIRONMENT_BANNER_FOREGROUND
* ENVIRONMENT_BANNER_BACKGROUND
* ENVIRONMENT_BANNER_ENABLED

The ENVIRONMENT_BANNER_ENABLED is only needed to force the banner to
be displayed when it otherwise wouldn't be (such as on a production
server).

Added "byebug" as a development dependency to help troubleshoot
tests.

https://issues.umd.edu/browse/LIBSEARCH-120